server/pyproject.toml
[tool.poetry]
name = "emissions-bot-server"
version = "0.0.1"
description = ""
authors = ["Craig Fiedorek <rcraigfiedorek@gmail.com>"]
packages = [
    { include = "api" },
    { include = "db" },
]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.1"
requests = "^2.28.1"
apiflask = "1.1.2"
flask-sqlalchemy = "^3.0.2"
gunicorn = "^20.1.0"
cleanco = "2.2"
google-cloud-secret-manager = "^2.12.6"
flask-cors = "^3.0.10"
pint = "^0.20.1"
pylcs = "^0.0.7"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
jupyter = "^1.0.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
preview = true
